LoremIpsum.Accounts (Lorem Ipsum v0.0.0)

The Accounts context.

Functions

from_token(token)

user_from_claim(claims, token)

@spec user_from_claim(claims :: map(), token :: binary()) ::
  {:ok, LoremIpsum.Accounts.User.t()}

Retrieves, inserts, or updates a user record corresponding to the "sub" claim of a validated token.

If the user record does not exist, it makes a call to Auth0 and requests the corresponding user profile data, then creates the record.

If the user record exists and has the email validated, it loads the record from the database.

If the user record exists but the email is not validated, it makes a call to Auth0 and requests the corresponding user profile data. If the requested data indicates that the email has been verified, the record in the database is updated.

Examples

iex> token = "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI..."
iex> claims = %{
...>   "sub" => "auth0|664ad728cddc69d8a11f0368"
...> }
iex> %{elixir_module}.Accounts.user_from_claim(claims, token)
{:ok, %User{}}

iex> token = "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI..."
iex> claims = %{
...>   "sub" => "auth0|inexistent"
...> }
iex> %{elixir_module}.Accounts.user_from_claim(claims, token)
{:ok, {:error, %{
  code: 401,
  message: "unauthenticated",
  detail: "some error"
}}}