LoremIpsum.Assistant (Lorem Ipsum v0.0.0)

The Assistant context handles operations related to user conversations with the OpenAI API.

This module provides functionality for managing conversations, including creating, retrieving, continuing, and deleting them. It ensures that all operations are scoped to the authenticated user, maintaining proper access control.

This module acts as the bridge between user interactions and the OpenAI API, while also persisting conversation data securely in the database.

Functions

continue_conversation(attrs, current_user)

@spec continue_conversation(
  attrs :: map(),
  current_user :: LoremIpsum.Accounts.User.t()
) ::
  {:ok, LoremIpsum.Assistant.Conversation.t()}
  | {:error, Ecto.Changeset.t() | :not_found | :not_owner}

Sends a request to the OpenAI API to process all conversation messages, including the new ones. If the request is successful, it saves the messages to the conversation and returns the updated conversation.

Examples

iex> continue_conversation(attrs, %User{})
{:ok, %Conversation{}}

iex> get_conversation(attrs, %User{})
{:error, %Ecto.Changeset{}}

iex> continue_conversation(attrs, %User{})
{:error, :not_found}

iex> continue_conversation(attrs, %User{})
{:error, :not_owner}

create_conversation(attrs, current_user)

@spec create_conversation(
  attrs :: map(),
  current_user :: LoremIpsum.Accounts.User.t()
) ::
  {:ok, LoremIpsum.Assistant.Conversation.t()} | {:error, Ecto.Changeset.t()}

Sends a request to the OpenAI API to process all conversation messages. If the request is successful, it saves the messages to the conversation and returns the newly created conversation.

Examples

iex> create_conversation(attrs, %User{})
{:ok, %Conversation{}}

iex> create_conversation(attrs, %User{})
{:error, %Ecto.Changeset{}}

iex> continue_conversation(attrs, %User{})
{:error, :not_found}

iex> continue_conversation(attrs, %User{})
{:error, :not_owner}

delete_conversation(attrs, current_user)

@spec delete_conversation(
  attrs :: map(),
  current_user :: LoremIpsum.Accounts.User.t()
) ::
  {:ok, LoremIpsum.Assistant.Conversation.t()}
  | {:error, :not_found | :not_owner}

Deletes a conversation if it belongs to the given user.

Examples

iex> delete_conversation(attrs, %User{})
{:ok, %Conversation{}}

iex> delete_conversation(attrs, %User{})
{:error, :not_found}

iex> delete_conversation(attrs, %User{})
{:error, :not_owner}

get_conversation(attrs, current_user)

@spec get_conversation(attrs :: map(), current_user :: LoremIpsum.Accounts.User.t()) ::
  {:ok, LoremIpsum.Assistant.Conversation.t()}
  | {:error, Ecto.Changeset.t() | :not_found | :not_owner}

Retrieves a single conversation with all its messages if it belongs to the given user.

Examples

iex> get_conversation(attrs, %User{})
{:ok, %Conversation{}}

iex> get_conversation(attrs, %User{})
{:error, %Ecto.Changeset{}}

iex> get_conversation(attrs, %User{})
{:error, :not_found}

iex> get_conversation(attrs, %User{})
{:error, :not_owner}

list_conversations(attrs, current_user)

@spec list_conversations(attrs :: map(), current_user :: LoremIpsum.Accounts.User.t()) ::
  {:ok,
   %{
     records: [LoremIpsum.Assistant.Conversation.t()],
     count: pos_integer(),
     total_count: pos_integer(),
     page: non_neg_integer(),
     total_pages: non_neg_integer(),
     query: %{
       limit: pos_integer(),
       offset: non_neg_integer(),
       field: atom(),
       order: :asc | :desc
     }
   }}

Returns the list of conversations belonging to the given user.

Examples

iex> list_conversations(attrs, %User{})
{:ok, %{
  records: [
    %Conversation{},
    ...
  ],
  count: 15,
  total_count: 47,
  page: 1,
  total_pages: 5,
  query: %{limit: 15, offset: 0, field: :name, order: :asc}
}}