LoremIpsum.Helper (Lorem Ipsum v0.0.0)

Module for handling standard pagination, and changeset logic tools.

Provides functions to retrieve standarized paginated resource lists from the database, and traverse check for specific fields errors in changesets.

Functions

errors_on?(changeset)

@spec errors_on?(changeset :: Ecto.Changeset.t()) :: boolean()

Returns boolean indicating if the changeset is valid, traversing the changset looking for any errors.

Example

iex> changeset = %Changeset(valid?: false)
iex> Helper.errors_on?(changeset)
true

iex> changeset = %Changeset(valid?: true)
iex> Helper.errors_on?(changeset)
false

errors_on?(changeset, field)

@spec errors_on?(changeset :: Ecto.Changeset.t(), field :: atom() | [atom()] | nil) ::
  boolean()

Returns boolean indicating if the changeset is valid, traversing the changset errors and checking only for specified field errors.

Returns true if the changset have an error attributed to the given changeset field or the given list of fields.

Returns false if the changeset does not have an error attributed to the given changeset field or the given list of fields.

Example

iex> changeset = %Changeset(errors: [field: {"is invalid", []}])
iex> Helper.errors_on?(changeset, :field)
true

iex> changeset = %Changeset(errors: [])
iex> Helper.errors_on?(changeset, :field)
false

iex> changeset = %Changeset(errors: [field_02: {"is invalid", []}])
iex> Helper.errors_on?(changeset, [:field_01, :field_02, :field_03])
true

iex> changeset = %Changeset(errors: [field_04: {"is invalid", []}])
iex> Helper.errors_on?(changeset, [:field_01, :field_02, :field_03])
false

paginate(query, attrs \\ %{})

@spec paginate(query :: module() | Ecto.Query.t(), attrs :: map()) :: map()

Gets a paginated list of resources from the database based on the given query.

The attrs parameter must be a map of string type keys which:

"limit": Number of records to return in a page.

If the key does not exist it will be set to 15 by default.

If the key value does not represent an integer it will be set to 15 by default.

If the key value is > 250 it will be set to 250 by default.

If the key value is < 1 it will be set to 1 by default.

"offset": Number of pages to skip.

If the key does not exist it will be set to 0 by default.

If the key value does not represent an integer it will be set to 0 by default.

If the key value is < 0 it will be set to 0.

"field": Specifies which field to use when sorting the results.

If the key does not exist it will be set to "inserted_at" by default.

If the key value does not represent an existing schema field, it will be set to "inserted_at" by default.

"order": Determines the sorting direction of the results. Accepts "asc" for ascending or "desc" for descending order.

If the key does not exist it will be set to "asc" by default.

If the key value is not "asc" or "desc" it will be set to "asc" by default.

Example

iex> Helper.paginate(User)
%{
  records: [
    %User{},
    %User{},
    %User{}
  ],
  count: 3,
  page: 1,
  total_count: 3,
  total_pages: 1,
  query: %{limit: 15, offset: 0, order: "asc", field: "inserted_at"}
}