LoremIpsumWeb.ConversationController (Lorem Ipsum v0.0.0)

Conversation Operations controller.

Functions

continue_conversation(conn, params)

@spec continue_conversation(
  conn :: Plug.Conn.t(),
  params :: %{optional(binary()) => any()}
) :: Plug.Conn.t()

Send new message(s) within an existing conversation.

Continue a specified conversation by adding new message(s) and requesting a response. The updated conversation, including the assistant's response, is saved in the database. This endpoint returns only the assistant's latest response message.

create_conversation(conn, params)

@spec create_conversation(
  conn :: Plug.Conn.t(),
  params :: %{optional(binary()) => any()}
) :: Plug.Conn.t()

Start a new conversation with the assistant.

Initiate a new conversation with the assistant by sending the first message(s). The conversation, including the assistant's response, is saved in the database. This endpoint returns only the assistant's latest response message.

delete_conversation(conn, params)

@spec delete_conversation(
  conn :: Plug.Conn.t(),
  params :: %{optional(binary()) => any()}
) :: Plug.Conn.t()

Delete a user's single conversation.

Remove a single conversation with the assistant that belongs to the active session user.

get_conversation(conn, params)

@spec get_conversation(
  conn :: Plug.Conn.t(),
  params :: %{optional(binary()) => any()}
) :: Plug.Conn.t()

Get a user's single conversation.

Retrieve a single conversation with the assistant that belongs to the active session user.

list_conversations(conn, params)

@spec list_conversations(
  conn :: Plug.Conn.t(),
  params :: %{optional(binary()) => any()}
) :: Plug.Conn.t()

Get the user's conversations list.

Retrieve a paginated list of the active session user's conversations with the assistant.

open_api_operation(action)

@spec open_api_operation(atom()) :: OpenApiSpex.Operation.t()

shared_security()

shared_tags()