LoremIpsumWeb.ErrorJSON (Lorem Ipsum v0.0.0)

This module is invoked by your endpoint in case of errors on JSON requests.

See config/config.exs.

Functions

error(map)

@spec error(%{changeset: changeset :: Ecto.Changeset.t()}) :: %{error: map()}

@spec error(%{message: message :: binary()}) :: %{error: binary()}

Standard changeset errors render.

Renders an error response body containing all traversed errors present in the changeset, including the embedded changesets.

Example

iex> LoremIpsumWeb.ErrorJSON.error(%{changeset: %Changeset{}})
%{error: %{
  field_01: ["is invalid"],
  field_02: ["can't be blank"]
}}

render(template, assigns)

@spec render(template :: binary(), _assigns :: term()) :: %{error: binary()}

Default HTTP errors render.

Renders an error response body containing the Phoenix default HTTP error message.

Example

iex> LoremIpsumWeb.ErrorJSON.render("404.json", %{})
%{error: "Not Found"}

iex> LoremIpsumWeb.ErrorJSON.render("500.json", %{})
%{error: "Internal Server Error"}