LoremIpsumWeb.ExDocController (Lorem Ipsum v0.0.0)

ExDoc documentation pages controller.

Functions

cover(conn, params)

@spec cover(conn :: Plug.Conn.t(), _params :: map()) :: conn :: Plug.Conn.t()

For a request specifiying coverage, sends the coverage report page.

handle(conn, params)

@spec handle(conn :: Plug.Conn.t(), _params :: map()) :: conn :: Plug.Conn.t()

If the requested specific page does not exists, sends the not-found error page.

index(conn, params)

@spec index(conn :: Plug.Conn.t(), _params :: map()) :: conn :: Plug.Conn.t()

For a request with no specified page, sends the index page.